cleanup unused commented out code
