+++ src/scenes/categoryDetails/CategoryDetails.tsx
@@ -1,9 +1,4 @@
-import React, { useState } from 'react';
-import { Box, Typography } from '@mui/material';
-import SearchBar from '../../components/SearchBar';
-import LanguageDropdown from '../../components/LanguageDropdown';
-import { Data } from '../../data/Data';
-import CategoryList from '../../components/CategoryList';
+import React from 'react';
 import { useLocation, useParams } from 'react-router-dom';
 
 const CategoryDetails = () => {
